The checkout page is working again and all the code has been commented out at the moment. Everything is in the useForm hook now.
